+++ tiledb/sm/serialization/query.cc
@@ -32,6 +32,7 @@
  */
 
 #include "tiledb/sm/serialization/query.h"
+#include "tiledb/common/heap_memory.h"
 #include "tiledb/common/logger.h"
 #include "tiledb/sm/array/array.h"
 #include "tiledb/sm/buffer/buffer_list.h"
@@ -976,8 +977,8 @@ Status query_from_capnp(
         void* subarray = nullptr;
         RETURN_NOT_OK(
             utils::deserialize_subarray(subarray_reader, schema, &subarray));
-        RETURN_NOT_OK_ELSE(query->set_subarray(subarray), std::free(subarray));
-        std::free(subarray);
+        RETURN_NOT_OK_ELSE(query->set_subarray(subarray), tdb_free(subarray));
+        tdb_free(subarray);
       }
 
       // Subarray